Added new instruction emulating functions

+++ sim/regaccess-riscv.c
@@ -82,9 +82,9 @@ void reg_set_riscv(Engine *E, State *S, uint8_t n, uint32_t data)
 
 /* RV32F floating point register access */
 tuck
-uint32_t freg_read_riscv(Engine *E, State *S, uint8_t n)
+uint64_t freg_read_riscv(Engine *E, State *S, uint8_t n)
 {
-	uint32_t data;
+	uint64_t data;
 
 	if (n < RF32F_fMAX)
 	{
@@ -95,7 +95,7 @@ uint32_t freg_read_riscv(Engine *E, State *S, uint8_t n)
 }
 
 tuck
-void freg_set_riscv(Engine *E, State *S, uint8_t n, uint32_t data)
+void freg_set_riscv(Engine *E, State *S, uint8_t n, uint64_t data)
 {
 	if (n < RF32F_fMAX)
 	{
